+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.77.0 (Unreleased)
 BUG FIXES:
 * iam: fix access bindings deletion
+* ydb: fixed panic if database was removed outside of terraform
 
 ENHANCEMENTS:
 * mdb: add `sqlcollation` attribute to `yandex_mdb_sqlserver_cluster` resource and data source